Refactor tests

+++ src/test/flight-test.spec.ts
@@ -9,6 +9,8 @@ import { RegisterDTO } from '../auth/dto/register.dto';
 import { LoginDTO } from '../auth/dto/login.dto';
 import { Auth } from '../auth/entity/auth.entity';
 import { FlightDTO } from '../flight/dto/fligth.dto';
+import { FlightResponseDTO } from '../flight/dto/fligth-response.dto';
+import { plainToInstance } from 'class-transformer';
 
 describe('FlightController', () => {
   let app: INestApplication;
@@ -16,7 +18,7 @@ describe('FlightController', () => {
   let flightRepository: Repository<Flight>;
   let authRepository: Repository<Auth>;
   let accessToken = null;
-  let flightDB: Flight;
+  let flight: FlightResponseDTO;
 
     beforeAll(async () => {
         const moduleFixture: TestingModule = await Test.createTestingModule({
@@ -29,7 +31,7 @@ describe('FlightController', () => {
         authRepository = moduleFixture.get<Repository<Auth>>(getRepositoryToken(Auth));
 
         const registerDTO = new RegisterDTO();
-        registerDTO.email = "ADMIN1@GMAIL.COM";
+        registerDTO.email = "ADMIN2@GMAIL.COM";
         registerDTO.password = "12345";
         await request(app.getHttpServer())
             .post('/auth/register')
@@ -38,7 +40,7 @@ describe('FlightController', () => {
         
 
         const loginDTO = new LoginDTO();
-        loginDTO.email = "ADMIN1@GMAIL.COM";
+        loginDTO.email = "ADMIN2@GMAIL.COM";
         loginDTO.password = "12345";
         const responseLogin = await request(app.getHttpServer())
             .post('/auth/login')
@@ -46,61 +48,21 @@ describe('FlightController', () => {
             .expect(200)
         accessToken = responseLogin.body.token;
 
-        flightDB = new Flight();
-        flightDB.airline = "Test Airline";
-        flightDB.origin = "Test Origin";
-        flightDB.destiny = "Test Destiny";
-        flightDB.departure = new Date();
-        flightDB.layovers = true;
-        flightDB.price = 15000;
-        flightDB = await flightRepository.save(flightDB)
-
     }, timeout);
 
 
     afterAll(async () => {
-      await flightRepository.delete({});
-      await authRepository.delete({})
+      await flightRepository.delete({id: flight.id});
+      await authRepository.delete({email: "ADMIN2@GMAIL.COM"})
       await app.close();
         
     }, timeout);
 
-  it('/flight (GET)', async () => {
-    const response = await request(app.getHttpServer())
-        .get('/flight')
-        .set('Authorization', `Bearer ${accessToken}`)
-        .expect(200)
-
-        expect(Array.isArray(response.body)).toBe(true);
-        expect(response.body.length).toBeGreaterThan(0);
-
-        expect(response.body[0].id).toBeDefined();
-        expect(response.body[0].airline).toBe(flightDB.airline);
-        expect(response.body[0].origin).toBe(flightDB.origin);
-        expect(response.body[0].destiny).toBe(flightDB.destiny);
-        expect(response.body[0].layovers).toBe(flightDB.layovers);
-        expect(response.body[0].price).toBe(flightDB.price);
-  }, timeout);
-
-  it('/flight/:id (GET)', async () => {
-    const response = await request(app.getHttpServer())
-        .get('/flight/'+ flightDB.id)
-        .set('Authorization', `Bearer ${accessToken}`)
-        .expect(200)
-
-        expect(response.body.id).toBeDefined();
-        expect(response.body.airline).toBe(flightDB.airline);
-        expect(response.body.origin).toBe(flightDB.origin);
-        expect(response.body.destiny).toBe(flightDB.destiny);
-        expect(response.body.layovers).toBe(flightDB.layovers);
-        expect(response.body.price).toBe(flightDB.price);
-  }, timeout);
-
   it('/flight (POST)', async () => {
     const flightDTO = new FlightDTO();
-    flightDTO.airline = "Test Airline 2";
-    flightDTO.origin = "Test Origin 2";
-    flightDTO.destiny = "Test Destiny 2";
+    flightDTO.airline = "Test Airline";
+    flightDTO.origin = "Test Origin";
+    flightDTO.destiny = "Test Destiny";
     flightDTO.departure = new Date();
     flightDTO.layovers = true;
     flightDTO.price = 25000;
@@ -109,21 +71,51 @@ describe('FlightController', () => {
         .set('Authorization', `Bearer ${accessToken}`)
         .send(flightDTO)
         .expect(201)
+    flight = plainToInstance(FlightResponseDTO, response.body, { excludeExtraneousValues: true });
+    expect(flight.id).toBeDefined();
+    expect(flight.airline).toBe(flightDTO.airline);
+    expect(flight.origin).toBe(flightDTO.origin);
+    expect(flight.destiny).toBe(flightDTO.destiny);
+    expect(flight.layovers).toBe(flightDTO.layovers);
+    expect(flight.price).toBe(flightDTO.price);
+  }, timeout);
+
+  it('/flight (GET)', async () => {
+    const response = await request(app.getHttpServer())
+        .get('/flight')
+        .set('Authorization', `Bearer ${accessToken}`)
+        .expect(200)
+  
+    expect(Array.isArray(response.body)).toBe(true);
+    expect(response.body.length).toBeGreaterThan(0);
+
+    expect(response.body[0].id).toBeDefined();
+    expect(response.body[0].airline).toBeDefined();
+    expect(response.body[0].origin).toBeDefined();
+    expect(response.body[0].destiny).toBeDefined();
+    expect(response.body[0].layovers).toBeDefined();
+    expect(response.body[0].price).toBeDefined();
+  }, timeout);
+
+  it('/flight/:id (GET)', async () => {
+    const response = await request(app.getHttpServer())
+        .get('/flight/'+ flight.id)
+        .set('Authorization', `Bearer ${accessToken}`)
+        .expect(200)
 
         expect(response.body.id).toBeDefined();
-        expect(response.body.airline).toBe(flightDTO.airline);
-        expect(response.body.origin).toBe(flightDTO.origin);
-        expect(response.body.destiny).toBe(flightDTO.destiny);
-        expect(response.body.layovers).toBe(flightDTO.layovers);
-        expect(response.body.price).toBe(flightDTO.price);
+        expect(response.body.airline).toBe(flight.airline);
+        expect(response.body.origin).toBe(flight.origin);
+        expect(response.body.destiny).toBe(flight.destiny);
+        expect(response.body.layovers).toBe(flight.layovers);
+        expect(response.body.price).toBe(flight.price);
   }, timeout);
 
   it('/flight (UPDATE)', async () => {
-    const flight = await flightRepository.save(flightDB);
     const flightDTO = new FlightDTO();
-    flightDTO.airline = "Test Airline 3";
-    flightDTO.origin = "Test Origin 3";
-    flightDTO.destiny = "Test Destiny 3";
+    flightDTO.airline = "Test Airline 2";
+    flightDTO.origin = "Test Origin 2";
+    flightDTO.destiny = "Test Destiny 2";
     flightDTO.departure = new Date();
     flightDTO.layovers = false;
     flightDTO.price = 45000;
@@ -143,7 +135,7 @@ describe('FlightController', () => {
 
   it('/flight (DELETE)', async () => {
     await request(app.getHttpServer())
-        .delete("/flight/" + flightDB.id)
+        .delete("/flight/" + flight.id)
         .set('Authorization', `Bearer ${accessToken}`)
         .expect(200)
   }, timeout);
